// SmartValley.FrontEnd/src/app/services/project/project-area-type.enum.ts
export enum ProjectAreaTypeEnum {
  Art = 0,
  ArtificialIntelligence = 1,
  Banking = 2,
  BigData = 3,
  BusinessServices = 4,
  CasinoAndGambling = 5,
  Charity = 6,
  Communication = 7,
  Cryptocurrency = 8,
  Education = 9,
  Electronics = 10,
  Energy = 11,
  Entertainment = 12,
  Health = 13,
  Infrastructure = 14,
  Internet = 15,
  Investment = 16,
  Legal = 17,
  Manufacturing = 18,
  Media = 19,
  Other = 20,
  Platform = 21,
  RealEstate = 22,
  Retail = 23,
  SmartContract = 24,
  Software = 25,
  Sports = 26,
  Tourism = 27,
  VirtualReality = 28
}